Move UserProfilePage to its own folder
